moved loop-params to @/types/
