+++ src/types/loop-params.ts
